requirements.txt
Flask>=2.3.0
websocket-client>=1.6.0
requests-toolbelt>=1.0.0
Pillow>=10.0.0